F [#4ubrLq42]: Extracto Coin class from treasure module
The Coin class is not necessarily related to the treasure module, so I'm extracting it so that it is in the root level of the lib folder
